(BSR)[API] refactor: move common individual routes to product swagger
